repo_fetcher/requirements.txt
requests  
chardet  
PyYAML  
configparser  
